Move a bunch of functions to use the new proc macro
