+++ blackjack_engine/src/lua_engine/lua_stdlib/lua_mesh_library.rs
@@ -19,83 +19,6 @@ pub fn load(lua: &Lua) -> anyhow::Result<()> {
     let ops = lua.create_table()?;
     globals.set("Ops", ops.clone())?;
 
-    lua_fn!(lua, ops, "chamfer", |vertices: SelectionExpression,
-                                  amount: f32,
-                                  mesh: AnyUserData|
-     -> () {
-        let mesh = mesh.borrow_mut::<HalfEdgeMesh>()?;
-        mesh.write_connectivity().clear_debug();
-        let verts = mesh
-            .resolve_vertex_selection_full(&vertices)
-            .map_lua_err()?;
-        for v in verts {
-            crate::mesh::halfedge::edit_ops::chamfer_vertex(
-                &mut mesh.write_connectivity(),
-                &mut mesh.write_positions(),
-                v,
-                amount,
-            )
-            .map_lua_err()?;
-        }
-        Ok(())
-    });
-
-    lua_fn!(lua, ops, "bevel", |edges: SelectionExpression,
-                                amount: f32,
-                                mesh: AnyUserData|
-     -> () {
-        let result = mesh.borrow_mut::<HalfEdgeMesh>()?;
-        {
-            let edges = result
-                .resolve_halfedge_selection_full(&edges)
-                .map_lua_err()?;
-            crate::mesh::halfedge::edit_ops::bevel_edges(
-                &mut result.write_connectivity(),
-                &mut result.write_positions(),
-                &edges,
-                amount,
-            )
-            .map_lua_err()?;
-        }
-        Ok(())
-    });
-
-    lua_fn!(lua, ops, "extrude", |faces: SelectionExpression,
-                                  amount: f32,
-                                  mesh: AnyUserData|
-     -> () {
-        let result = mesh.borrow_mut::<HalfEdgeMesh>()?;
-        {
-            let faces = result.resolve_face_selection_full(&faces).map_lua_err()?;
-            crate::mesh::halfedge::edit_ops::extrude_faces(
-                &mut result.write_connectivity(),
-                &mut result.write_positions(),
-                &faces,
-                amount,
-            )
-            .map_lua_err()?;
-        }
-        Ok(())
-    });
-
-    lua_fn!(lua, ops, "merge", |a: AnyUserData, b: AnyUserData| -> () {
-        let mut a = a.borrow_mut::<HalfEdgeMesh>()?;
-        let b = b.borrow::<HalfEdgeMesh>()?;
-        a.merge_with(&b);
-        Ok(())
-    });
-
-    lua_fn!(lua, ops, "subdivide", |mesh: AnyUserData,
-                                    iterations: usize,
-                                    catmull_clark: bool|
-     -> HalfEdgeMesh {
-        let mesh = &mesh.borrow::<HalfEdgeMesh>()?;
-        let new_mesh = CompactMesh::<false>::from_halfedge(mesh).map_lua_err()?;
-        Ok(new_mesh
-            .subdivide_multi(iterations, catmull_clark)
-            .to_halfedge())
-    });
-
     lua_fn!(lua, ops, "set_smooth_normals", |mesh: AnyUserData| -> () {
         let mut mesh = mesh.borrow_mut::<HalfEdgeMesh>()?;
         crate::mesh::halfedge::edit_ops::set_smooth_normals(&mut mesh).map_lua_err()?;
